Dockerize (#22)
* Adds Makefile for local development

* Fix chat completions example request

* Added delete example

* Dockerize model-runner

* WIP Run container with host access to socket

* Dockerize model-runner

* WIP Run container with host access to socket

* Debugging

* Run in Docker container with TCP port access

* mounted model storage

* - Remove duplication in .gitignore
- Do not use alpine in builder image
- NVIDIA seems to use Ubuntu in all of their CDI docs and produces Ubuntu tags for nvidia/cuda but not Debian. So use Ubuntu for our final image
For more details: https://github.com/docker/model-runner/pull/22

* - Add MODELS_PATH environment variable to configure model storage location
- Default to $HOME/.docker/models when MODELS_PATH is not set
- Update Docker container to use /models as the default storage path
- Update Makefile to pass MODELS_PATH to container
- Update Dockerfile to create and set permissions for /models directory

This change allows users to:
- Override the model storage location via MODELS_PATH
- Maintain backward compatibility with default $HOME/.docker/models path
- Use a more idiomatic folder for /models

* Removes unneeded logs

+++ main.go
@@ -32,16 +32,28 @@ func main() {
 		log.Fatalf("Failed to get user home directory: %v", err)
 	}
 
+	modelPath := os.Getenv("MODELS_PATH")
+	if modelPath == "" {
+		modelPath = filepath.Join(userHomeDir, ".docker", "models")
+	}
+
 	modelManager := models.NewManager(log, models.ClientConfig{
-		StoreRootPath: filepath.Join(userHomeDir, ".docker", "models"),
+		StoreRootPath: modelPath,
 		Logger:        log.WithFields(logrus.Fields{"component": "model-manager"}),
 	})
+
+	llamaServerPath := os.Getenv("LLAMA_SERVER_PATH")
+	if llamaServerPath == "" {
+		llamaServerPath = "/Applications/Docker.app/Contents/Resources/bin"
+	}
+
+	log.Infof("LLAMA_SERVER_PATH: %s", llamaServerPath)
 
 	llamaCppBackend, err := llamacpp.New(
 		log,
 		modelManager,
 		log.WithFields(logrus.Fields{"component": "llama.cpp"}),
-		"/Applications/Docker.app/Contents/Resources/bin",
+		llamaServerPath,
 		func() string { wd, _ := os.Getwd(); return wd }(),
 	)
 	if err != nil {
@@ -64,22 +76,34 @@ func main() {
 		router.Handle(route, scheduler)
 	}
 
-	if err := os.Remove(sockName); err != nil {
-		if !os.IsNotExist(err) {
-			log.Fatalf("Failed to remove existing socket: %v", err)
-		}
-	}
-	ln, err := net.ListenUnix("unix", &net.UnixAddr{Name: sockName, Net: "unix"})
-	if err != nil {
-		log.Fatalf("Failed to listen on socket: %v", err)
-	}
-
 	server := &http.Server{Handler: router}
 	serverErrors := make(chan error, 1)
-	go func() {
-		serverErrors <- server.Serve(ln)
-	}()
-	defer server.Close()
+
+	// Check if we should use TCP port instead of Unix socket
+	tcpPort := os.Getenv("MODEL_RUNNER_PORT")
+	if tcpPort != "" {
+		// Use TCP port
+		addr := ":" + tcpPort
+		log.Infof("Listening on TCP port %s", tcpPort)
+		server.Addr = addr
+		go func() {
+			serverErrors <- server.ListenAndServe()
+		}()
+	} else {
+		// Use Unix socket
+		if err := os.Remove(sockName); err != nil {
+			if !os.IsNotExist(err) {
+				log.Fatalf("Failed to remove existing socket: %v", err)
+			}
+		}
+		ln, err := net.ListenUnix("unix", &net.UnixAddr{Name: sockName, Net: "unix"})
+		if err != nil {
+			log.Fatalf("Failed to listen on socket: %v", err)
+		}
+		go func() {
+			serverErrors <- server.Serve(ln)
+		}()
+	}
 
 	schedulerErrors := make(chan error, 1)
 	go func() {
